committed Numbers have life; they're not just symbols on paper

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -8,7 +8,7 @@
 
 void print_number(int n)
 {
-	unsigned long int a;
+	unsigned int a;
 
 	a = n;
 	if (n < 0)
